Move pkg/util/configz to k8s.io/component-base
