+++ cmd/kube-controller-manager/app/BUILD
@@ -85,7 +85,6 @@ go_library(
         "//pkg/quota/v1/generic:go_default_library",
         "//pkg/quota/v1/install:go_default_library",
         "//pkg/serviceaccount:go_default_library",
-        "//pkg/util/configz:go_default_library",
         "//pkg/volume:go_default_library",
         "//pkg/volume/awsebs:go_default_library",
         "//pkg/volume/azure_dd:go_default_library",
@@ -139,6 +138,7 @@ go_library(
         "//staging/src/k8s.io/cloud-provider:go_default_library",
         "//staging/src/k8s.io/component-base/cli/flag:go_default_library",
         "//staging/src/k8s.io/component-base/cli/globalflag:go_default_library",
+        "//staging/src/k8s.io/component-base/configz:go_default_library",
         "//staging/src/k8s.io/component-base/featuregate:go_default_library",
         "//staging/src/k8s.io/component-base/metrics/prometheus/ratelimiter:go_default_library",
         "//staging/src/k8s.io/component-base/term:go_default_library",
