+++ cmd/kube-scheduler/app/testing/BUILD
@@ -9,10 +9,10 @@ go_library(
         "//cmd/kube-scheduler/app:go_default_library",
         "//cmd/kube-scheduler/app/config:go_default_library",
         "//cmd/kube-scheduler/app/options:go_default_library",
-        "//pkg/util/configz:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//staging/src/k8s.io/client-go/rest:go_default_library",
+        "//staging/src/k8s.io/component-base/configz:go_default_library",
         "//vendor/github.com/spf13/pflag:go_default_library",
     ],
 )
