+++ pkg/util/BUILD
@@ -14,7 +14,6 @@ filegroup(
         "//pkg/util/async:all-srcs",
         "//pkg/util/bandwidth:all-srcs",
         "//pkg/util/config:all-srcs",
-        "//pkg/util/configz:all-srcs",
         "//pkg/util/conntrack:all-srcs",
         "//pkg/util/coverage:all-srcs",
         "//pkg/util/ebtables:all-srcs",
