+++ staging/src/k8s.io/component-base/BUILD
@@ -13,6 +13,7 @@ filegroup(
         "//staging/src/k8s.io/component-base/cli/globalflag:all-srcs",
         "//staging/src/k8s.io/component-base/codec:all-srcs",
         "//staging/src/k8s.io/component-base/config:all-srcs",
+        "//staging/src/k8s.io/component-base/configz:all-srcs",
         "//staging/src/k8s.io/component-base/featuregate:all-srcs",
         "//staging/src/k8s.io/component-base/logs:all-srcs",
         "//staging/src/k8s.io/component-base/metrics:all-srcs",
